Fix compiler warnings and stale test assertions
- Elevator.cpp: cast Direction enum to double before multiplying
  (fixes -Wdeprecated-enum-float-conversion).
- MainMenu.cpp: replace deprecated tgui::Texture::load(sf::Texture)
  with loadFromPixelData() per TGUI 1.12 migration.
- test_main.cpp: update TimeWindowLayout assertions to match the
  current makeModernTimeWindowLayout values (clientHeight 80,
  tooltipWidth 298, tooltipBackgroundY 30). All tests now pass.

+++ source/Item/Elevator/Elevator.cpp
@@ -500,7 +500,7 @@ void Elevator::decideCarDestination(Car *car)
 		// lie behin the car a negative distance. We only serve queues that are somewhat ahead of
 		// the car.
 		double distance = q->floor - car->altitude;
-		distance *= car->direction;
+		distance *= static_cast<double>(car->direction);
 		if (distance < 0.5)
 			continue;
